+++ client/Player/components/PlayerView.tsx
@@ -6,25 +6,7 @@ import { PlayerViewState } from "../../../common/PlayerViewState";
 import { CustomStyles } from "./CustomStyles";
 import { PlayerViewCombatant } from "./PlayerViewCombatant";
 import { PlayerViewCombatantHeader } from "./PlayerViewCombatantHeader";
-
-interface PortraitModalProps {
-  imageURL: string;
-  caption: string;
-  onClose: () => void;
-}
-class PortraitModal extends React.Component<PortraitModalProps> {
-  public render() {
-    return (
-      <div
-        className="modal-blur combatant-portrait"
-        onClick={this.props.onClose}
-      >
-        <img className="combatant-portrait__image" src={this.props.imageURL} />
-        <div className="combatant-portrait__caption">{this.props.caption}</div>
-      </div>
-    );
-  }
-}
+import { PortraitModal } from "./PortraitModal";
 
 interface LocalState {
   showModal: boolean;
